refactor(#6) Dto 클래스 정리
