--- src/main/java/com/limito/payment/infrastructure/dto/request/ConfirmPaymentRequestV1.java
@@ -1,28 +1,0 @@
-package com.limito.payment.infrastructure.dto.request;
-
-import java.util.List;
-import java.util.UUID;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@ToString
-public class ConfirmPaymentRequestV1 {
-
-	private UUID orderId;
-
-	private String itemSummary;
-
-	private List<OrderItem> items;
-
-	private int totalPrice;
-}
