+++ src/main/java/com/limito/payment/infrastructure/dto/request/CreatePaymentRequestV1.java
@@ -1,4 +1,4 @@
-package com.limito.payment.application.dto.resquest;
+package com.limito.payment.infrastructure.dto.request;
 
 import java.util.List;
 
@@ -9,13 +9,10 @@ import jakarta.validation.constraints.Size;
 import lombok.Getter;
 import lombok.Setter;
 
-import com.limito.payment.presentation.dto.request.OrderItem;
-
 @Getter
 @Setter
 @NotBlank
-public class CreatePaymentResquestV1 {
-	@NotNull(message = "주문자 아이디는 필수입니다.")
+public class CreatePaymentRequestV1 {
 	private Long orderId;
 	@NotBlank(message = "상품 요약은 필수입니다.")
 	private String itemSummary;
@@ -24,4 +21,5 @@ public class CreatePaymentResquestV1 {
 	private int totalPrice;
 	@Size(min = 1, message = "결제 상품은 최소 1개 이상이어야 합니다.")
 	private List<OrderItem> items;
+
 }
