+++ src/main/java/com/limito/payment/presentation/PaymentInternalControllerV1.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.limito.payment.application.PaymentServiceV1;
-import com.limito.payment.presentation.dto.request.CreatePaymentRequestV1;
+import com.limito.payment.infrastructure.dto.request.CreatePaymentRequestV1;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -29,7 +29,7 @@ public class PaymentInternalControllerV1 {
 		@PathVariable String orderId,
 		@RequestBody CreatePaymentRequestV1 request) {
 		try {
-			paymentService.validatePaymentRequestUniqueness(UUID.fromString(orderId));
+			paymentService.validPaymentRequest(UUID.fromString(orderId), request);
 		} catch (Exception e) {
 			log.warn("주문 아이디 orderId={}에 대한 결제 요청 내역이 있습니다. {}", orderId, e.getMessage());
 			return ResponseEntity.status(HttpStatus.CONFLICT).body("중복 결제 불가");
